Fix Tinder DTO

+++ meet-all-api/src/main/java/fr/pinguet62/meetall/provider/tinder/dto/TinderLikeResponseDto.java
@@ -16,6 +16,6 @@ public class TinderLikeResponseDto {
     /***
      * When limit reached.
      */
-    private Integer rate_limited_until;
+    private Long rate_limited_until;
 
 }
